+++ CHANGELOG.md
@@ -8,6 +8,7 @@ dev
  ### Bug Fixes
 
  * ons-input: Add spellcheck to list of attributes passed to inner input. ([#2706](https://github.com/OnsenUI/OnsenUI/issues/2706)).
+ * ons.notification: Fix prompt callback not running when submitOnEnter is false. ([#2618](https://github.com/OnsenUI/OnsenUI/issues/2618)).
 
 2.10.10
 ---